Change `port` from uint16_t to uint32_t, to support VSOCK (#109)

+++ tests/event_stream_rpc_server_connection_test.c
@@ -251,7 +251,7 @@ static int s_fixture_setup_port0(struct aws_allocator *allocator, void *ctx) {
     test_data->listener = aws_event_stream_rpc_server_new_listener(allocator, &listener_options);
     ASSERT_NOT_NULL(test_data->listener);
 
-    uint16_t actual_port = aws_event_stream_rpc_server_listener_get_bound_port(test_data->listener);
+    uint32_t actual_port = aws_event_stream_rpc_server_listener_get_bound_port(test_data->listener);
     ASSERT_TRUE(actual_port > 0);
 
     test_data->allocator = allocator;
